Move triangulation objects to view package.
